--- src/nglui/easyviewer/__init__.py
@@ -1,3 +1,0 @@
-from ..site_utils import neuroglancer_url
-from .base import *
-from .ngl_components import *
